+++ src/shared/container/index.ts
@@ -5,9 +5,9 @@ import { IngredientsRepository } from '@modules/drinks/infra/prisma/IngredientsR
 import { IDrinksRepository } from '@modules/drinks/repositories/IDrinksRepository';
 import { DrinksRepository } from '@modules/drinks/infra/prisma/DrinksRepository';
 import { IUsersRepository } from '@modules/accounts/repositories/IUsersRepository';
-import { UsersRepository } from '@modules/accounts/infra/prisma/UsersRepository';
+import { UsersRepository } from '@modules/accounts/infra/prisma/repositories/UsersRepository';
 import { IUsersTokensRepository } from '@modules/accounts/repositories/IUsersTokensRepository';
-import { UsersTokensRepository } from '@modules/accounts/infra/prisma/UsersTokensRepository';
+import { UsersTokensRepository } from '@modules/accounts/infra/prisma/repositories/UsersTokensRepository';
 
 container.registerSingleton<IIngredientsRepository>('IngredientsRepository', IngredientsRepository);
