+++ src/shared/container/providers/date/IDateProvider.ts
@@ -6,7 +6,7 @@ interface IDateProvider {
 	addDays(days: number): Date;
 	addHours(hours: number): Date;
 	addMinutes(minutes: number): Date;
-	compareIfBefore(start_date: Date, end_date: Date): boolean;
+	compareIfBefore(dateA: Date, dateB: Date): boolean;
 }
 
 export { IDateProvider };
