+++ src/shared/container/providers/date/implementations/DayjsDateProvider.ts
@@ -40,8 +40,8 @@ class DayjsDateProvider implements IDateProvider {
 		return dayjs().add(minutes, 'minute').toDate();
 	}
 
-	compareIfBefore(start_date: Date, end_date: Date): boolean {
-		return dayjs(start_date).isBefore(end_date);
+	compareIfBefore(dateA: Date, dateB: Date): boolean {
+		return dayjs(dateA).isBefore(dateB);
 	}
 }
